# docs/swagger.yaml
basePath: /
definitions:
...
  model.DocLibraryDetailCreateReq:
    properties:
      content:
        type: string
      language:
        type: string
      library_id:
        type: integer
      parent_id:
        type: integer
      path:
        type: string
      title:
        type: string
      url:
        type: string
      version:
        type: string
    required:
    - library_id
    - title
    type: object
  model.DocLibraryDetailResp:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      language:
        type: string
      library_id:
        type: integer
      parent_id:
        type: integer
      path:
        type: string
      title:
        type: string
      updated_at:
        type: string
      url:
        type: string
      version:
        type: string
    type: object
  model.DocLibraryDetailTreeResp:
    properties:
      children:
        items:
          $ref: '#/definitions/model.DocLibraryDetailTreeResp'
        type: array
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      language:
        type: string
      library_id:
        type: integer
      parent_id:
        type: integer
      path:
        type: string
      title:
        type: string
      updated_at:
        type: string
      url:
        type: string
      version:
        type: string
    type: object
  model.DocLibraryDetailUpdateReq:
    properties:
      content:
        type: string
      id:
        type: integer
      language:
        type: string
      parent_id:
        type: integer
      path:
        type: string
      title:
        type: string
      url:
        type: string
      version:
        type: string
    required:
    - id
    - title
    type: object
...
  model.EssayCreateReq:
    properties:
      content:
        type: string
      draft:
        type: boolean
      images:
        items:
          type: string
        type: array
    type: object
  model.EssayResp:
    properties:
      content:
        type: string
      create_at:
        type: string
      draft:
        type: boolean
      id:
        type: integer
      images:
        items:
          type: string
        type: array
    type: object
  model.EssayUpdateReq:
    properties:
      content:
        type: string
      draft:
        type: boolean
      id:
        type: integer
      images:
        items:
          type: string
        type: array
    type: object
...
  model.HttpError:
    properties:
      code:
        type: integer
      msg:
        type: string
    type: object
  model.HttpSuccess:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  model.PageResult-model_DocLibraryDetailResp:
    properties:
      records:
        items:
          $ref: '#/definitions/model.DocLibraryDetailResp'
        type: array
      total:
        type: integer
    type: object
...
  model.PageResult-model_EssayResp:
    properties:
      records:
        items:
          $ref: '#/definitions/model.EssayResp'
        type: array
      total:
        type: integer
    type: object
...
host: localhost:13000
info:
  contact:
    email: fiber@swagger.io
    name: API Support
  description: This is a sample swagger for Fiber
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Fiber Example API
  version: "1.0"
paths:
...
  /api/v1/doc-library-detail/create:
    post:
      consumes:
      - application/json
      description: 创建新的文档库详情
      parameters:
      - description: 文档库详情创建请求
        in: body
        name: docLibraryDetail
        required: true
        schema:
          $ref: '#/definitions/model.DocLibraryDetailCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 创建文档库详情
      tags:
      - 文档库详情
  /api/v1/doc-library-detail/delete/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定文档库详情
      parameters:
      - description: 文档库详情 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 删除文档库详情
      tags:
      - 文档库详情
  /api/v1/doc-library-detail/page:
    get:
      consumes:
      - application/json
      description: 获取指定文档库下的文档库详情分页列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: size
        type: integer
      - description: 文档库 ID
        in: query
        name: library_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  $ref: '#/definitions/model.PageResult-model_DocLibraryDetailResp'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 获取文档库详情分页列表
      tags:
      - 文档库详情
  /api/v1/doc-library-detail/query/{id}:
    get:
      consumes:
      - application/json
      description: 获取指定文档库详情的信息
      parameters:
      - description: 文档库详情 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  $ref: '#/definitions/model.DocLibraryDetailResp'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 获取文档库详情
      tags:
      - 文档库详情
  /api/v1/doc-library-detail/tree:
    get:
      consumes:
      - application/json
      description: 获取指定文档库下的文档库详情树结构
      parameters:
      - description: 文档库 ID
        in: query
        name: library_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.DocLibraryDetailTreeResp'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 获取文档库详情树
      tags:
      - 文档库详情
  /api/v1/doc-library-detail/update/{id}:
    put:
      consumes:
      - application/json
      description: 更新指定文档库详情的信息
      parameters:
      - description: 文档库详情 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 文档库详情更新请求
        in: body
        name: docLibraryDetail
        required: true
        schema:
          $ref: '#/definitions/model.DocLibraryDetailUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 更新文档库详情
      tags:
      - 文档库详情
...
  /api/v1/essay/create:
    post:
      consumes:
      - application/json
      description: 创建一个新说说
      parameters:
      - description: 说创建请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/model.EssayCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  $ref: '#/definitions/model.EssayCreateReq'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 创建说说
      tags:
      - 说说
  /api/v1/essay/delete/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定说说
      parameters:
      - description: 说说ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  $ref: '#/definitions/model.EssayResp'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 删除说说
      tags:
      - 说说
  /api/v1/essay/list:
    get:
      consumes:
      - application/json
      description: 获取说说列表
      parameters:
      - default: 10
        description: 数量限制
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.EssayResp'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 获取说说列表
      tags:
      - 说说
  /api/v1/essay/page:
    get:
      consumes:
      - application/json
      description: 获取说说分页列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  $ref: '#/definitions/model.PageResult-model_EssayResp'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 获取说说分页列表
      tags:
      - 说说
  /api/v1/essay/update/{id}:
    put:
      consumes:
      - application/json
      description: 更新指定说说的信息
      parameters:
      - description: 说说ID
        in: path
        name: id
        required: true
        type: string
      - description: 说说更新请求
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/model.EssayUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.HttpSuccess'
            - properties:
                data:
                  $ref: '#/definitions/model.EssayUpdateReq'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.HttpError'
      summary: 更新说说
      tags:
      - 说说
...
swagger: "2.0"